removed unwanted files; renamed folder with typo;
