+++ packages/infrastructure/src/errors.ts
