--- packages/infratructure/test/integration/helpers/db.ts
@@ -1,83 +1,0 @@
-/**
- License
- --------------
- Copyright © 2021 Mojaloop Foundation
-
- The Mojaloop files are made available by the Mojaloop Foundation under the Apache License, Version 2.0 (the "License") and you may not use these files except in compliance with the License.
-
- You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, the Mojaloop files are distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
-
- Contributors
- --------------
- This is the official list (alphabetical ordering) of the Mojaloop project contributors for this file.
- Names of the original copyright holders (individuals or organizations)
- should be listed with a '' in the first column. People who have
- contributed from an organization can be listed under the organization
- that actually holds the copyright for their contributions (see the
- Gates Foundation organization for an example). Those individuals should have
- their names indented and be marked with a '-'. Email address can be added
- optionally within square brackets <email>.
-
- * Gates Foundation
- - Name Surname <name.surname@gatesfoundation.com>
-
- * Coil
- - Jason Bruwer <jason.bruwer@coil.com>
-
- * Crosslake
- - Pedro Sousa Barreto <pedrob@crosslaketech.com>
-
- * Gonçalo Garcia <goncalogarcia99@gmail.com>
- 
- * Arg Software
- - José Antunes <jose.antunes@arg.software>
- - Rui Rocha <rui.rocha@arg.software>
-
- --------------
- **/
-
-"use strict";
-
-import { MongoClient } from "mongodb";
-
-export enum MongoDbOperationEnum {
-    INSERT_ONE = 'insertOne',
-    DELETE_MANY = 'deleteMany'
-}
-
-interface IMongoDBQuery {
-    dbUrl: string;
-    dbName: string;
-    dbCollection: string;
-    operation: MongoDbOperationEnum;
-    query: object;
-    cb?: Function
-}
-
-export async function mongoQuery({
-    dbUrl,
-    dbName,
-    dbCollection,
-    operation,
-    query = {},
-    cb = Function
-}: IMongoDBQuery) {
-    const client = new MongoClient(dbUrl);
-    
-    try {
-        const database = client.db(dbName);
-
-        const collection = database.collection(dbCollection);
-
-        await collection[operation](query);
-
-        if(cb) {
-            cb();
-        }
-        
-      } finally {
-        await client.close();
-      }
-}
